chore: move getLocale to gatsby/utils folder
