+++ packages/gatsby-theme-newrelic/gatsby/on-route-update.js
@@ -1,5 +1,5 @@
 import trackViaTessen from '../src/utils/page-tracking/tessen';
-import getLocale from '../src/utils/getLocale';
+import getLocale from './utils/getLocale';
 
 const onRouteUpdate = ({ location, prevLocation }, themeOptions) => {
   trackViaTessen({ location, prevLocation }, themeOptions);
