+++ packages/gatsby-theme-newrelic/gatsby/utils/getLocale.js
@@ -1,0 +1,15 @@
+const { getI18nConfig } = require('../../src/utils/config');
+
+const getLocale = ({ location }, themeOptions) => {
+  const { locales, defaultLocale } = getI18nConfig(themeOptions);
+
+  const [, base] = location.pathname.split('/');
+  const locale =
+    locales
+      .filter((locale) => !locale.isDefault)
+      .find(({ urlPrefix }) => urlPrefix === `/${base}`) || defaultLocale;
+
+  return locale.locale;
+};
+
+module.exports = getLocale;
